+++ ui/goose2/src/features/skills/lib/skillsHelpers.ts
@@ -1,4 +1,13 @@
 import type { SkillInfo } from "../api/skills";
+import type { SkillCategory, SkillViewInfo } from "./skillCategories";
+
+export type SkillsFilter = "all" | "global" | `project:${string}`;
+
+export interface SkillsSection {
+  id: string;
+  title: string;
+  skills: SkillViewInfo[];
+}
 
 // Mirrors crates/goose/src/skills/mod.rs::validate_skill_name.
 // Keep in sync with the Rust rule.
@@ -47,6 +56,104 @@ export function compareSkillsByName(a: SkillInfo, b: SkillInfo) {
     a.name.localeCompare(b.name) ||
     a.path.localeCompare(b.path)
   );
+}
+
+export function filterSkills(
+  skills: SkillViewInfo[],
+  filters: {
+    search: string;
+    activeFilter: SkillsFilter;
+    selectedCategories: SkillCategory[];
+  },
+  getCategoryLabel: (category: SkillCategory) => string,
+): SkillViewInfo[] {
+  const searchTerm = filters.search.trim().toLowerCase();
+  return skills.filter((skill) => {
+    const matchesSearch =
+      searchTerm.length === 0 ||
+      skill.name.toLowerCase().includes(searchTerm) ||
+      skill.description.toLowerCase().includes(searchTerm) ||
+      skill.sourceLabel.toLowerCase().includes(searchTerm) ||
+      getCategoryLabel(skill.inferredCategory)
+        .toLowerCase()
+        .includes(searchTerm);
+
+    const matchesFilter =
+      filters.activeFilter === "all"
+        ? true
+        : filters.activeFilter === "global"
+          ? skill.sourceKind === "global"
+          : skill.projectLinks.some(
+              (project) => `project:${project.id}` === filters.activeFilter,
+            );
+
+    const matchesCategory =
+      filters.selectedCategories.length === 0 ||
+      filters.selectedCategories.includes(skill.inferredCategory);
+
+    return matchesSearch && matchesFilter && matchesCategory;
+  });
+}
+
+export function groupSkills(
+  filteredSkills: SkillViewInfo[],
+  activeFilter: SkillsFilter,
+  projectFilters: { id: string; name: string }[],
+  labels: { personalTitle: string; projectsFallback: string },
+): SkillsSection[] {
+  if (activeFilter === "global") {
+    return [
+      {
+        id: "personal",
+        title: labels.personalTitle,
+        skills: [...filteredSkills].sort(compareSkillsByName),
+      },
+    ];
+  }
+
+  if (activeFilter.startsWith("project:")) {
+    const projectId = activeFilter.slice("project:".length);
+    const projectName =
+      projectFilters.find((project) => project.id === projectId)?.name ??
+      labels.projectsFallback;
+
+    return [
+      {
+        id: activeFilter,
+        title: projectName,
+        skills: [...filteredSkills].sort(compareSkillsByName),
+      },
+    ];
+  }
+
+  const personalSkills = filteredSkills
+    .filter((skill) => skill.sourceKind === "global")
+    .sort(compareSkillsByName);
+
+  const projectSections = projectFilters
+    .map((project) => ({
+      id: `project:${project.id}`,
+      title: project.name,
+      skills: filteredSkills
+        .filter((skill) =>
+          skill.projectLinks.some((link) => link.id === project.id),
+        )
+        .sort(compareSkillsByName),
+    }))
+    .filter((section) => section.skills.length > 0);
+
+  return [
+    ...(personalSkills.length > 0
+      ? [
+          {
+            id: "personal",
+            title: labels.personalTitle,
+            skills: personalSkills,
+          },
+        ]
+      : []),
+    ...projectSections,
+  ];
 }
 
 export function downloadExport(json: string, filename: string) {
